front (albums - create, musics - index e create)

+++ app/Http/Controllers/AlbumController.php
@@ -60,6 +60,10 @@ class AlbumController extends Controller
 
     public function destroy(Album $album)
     {
+        if ((bool)$album->musics->count()) {
+            return redirect()->back()->with('error', "Não é possível deletar um álbum que possui música(s) cadastrada(s)!");
+        }
+
         Storage::delete($album->cover_image);
         $album->delete();
